requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-limiter==3.5.0
flask-caching==2.1.0
pydantic==2.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
pytest==7.4.3
